+++ skrub/tests/test_datetime_encoder.py
@@ -299,6 +299,7 @@ def test_all_outputs_choice(datetime_cols, params, all_outputs):
     enc = DatetimeEncoder(**params)
     res = enc.fit_transform(datetime_cols.datetime)
     assert enc.all_outputs_ == [f"when_{f}" for f in all_outputs]
+    assert enc.get_feature_names_out() == [f"when_{f}" for f in all_outputs]
     assert sbd.column_names(res) == [f"{f}" for f in enc.all_outputs_]
 
 
@@ -368,3 +369,11 @@ def test_error_checking_periodic_encoder(a_datetime_col):
 def test_error_dispatch(func):
     with pytest.raises(TypeError, match="Expecting a Pandas or Polars Series"):
         func(np.array([1]))
+
+
+def test_n_splines_default_value(df_module):
+    """Check that when `n_splines is None`, it defaults to the `period` value."""
+    period = 15
+    enc = _SplineEncoder(period=period)
+    result = enc.fit_transform(df_module.make_column("when", [20, 20, 20]))
+    assert sbd.column_names(result) == [f"when_spline_{i:02d}" for i in range(period)]