+++ CHANGES.rst
@@ -11,6 +11,48 @@ Ongoing Development
 
 New features
 ------------
+- :meth:`DataOp.skb.apply` now allows passing extra named arguments to the
+  estimator's methods through the parameters ``fit_kwargs``, ``predict_kwargs``
+  etc. :pr:`1642` by :user:`Jérôme Dockès <jeromedockes>`.
+- TableReport now displays the mean statistic for boolean columns.
+  :pr:`1647` by :user:`Abdelhakim Benechehab <abenechehab>`.
+- :meth:`DataOp.skb.get_vars` allows inspecting all the variables, or all the
+  named dataops, in a :class:`DataOp`. This lets us easily know what keys should
+  be present in the ``environment`` dictionary we pass to
+  :meth:`DataOp.skb.eval` or to :meth:`SkrubLearner.fit`,
+  :meth:`SkrubLearner.predict`, etc. .
+  :pr:`1646` by :user:`Jérôme Dockès <jeromedockes>`.
+- :meth:`DataOp.skb.iter_cv_splits` iterates over the training and testing
+  environments produced by a CV splitter -- similar to
+  :meth:`DataOp.skb.train_test_split` but for multiple cross-validation splits.
+  :pr:`1653` by :user:`Jérôme Dockès <jeromedockes>`.
+- :meth:`DataOp.skb.full_report` now accepts a new parameter, title, that is displayed
+  in the html report.
+  :pr:`1654` by :user:`Marie Sacksick <MarieSacksick>`.
+
+Changes
+-------
+- The :func: `tabular_pipeline` uses a :class:`SquashingScaler` instead of a
+  :class:`StandardScaler` for centering and scaling numerical features
+  when linear models are used.
+  :pr:`1644` by :user:`Simon Dierickx <dierickxsimon>`
+
+Bugfixes
+--------
+- Issues occurring when :meth:`DataOp.skb.apply` was passed a DataOp as the
+  estimator have been fixed in :pr:`1671` by :user:`Jérôme Dockès
+  <jeromedockes>`.
+- :class:`TableReport` could raise an error while trying to check if Polars
+  columns with some dtypes (lists, structs) are sorted. It would not indicate
+  Polars columns sorted in descending order. Fixed in :pr:`1673` by
+  :user:`Jérôme Dockès <jeromedockes>`.
+
+
+Release 0.6.2
+=============
+
+New features
+------------
 - The :meth:`DataOp.skb.full_report` now displays the time each node took to
   evaluate. :pr:`1596` by :user:`Jérôme Dockès <jeromedockes>`.
 
@@ -20,6 +62,16 @@ Changes
   :func:`datasets.get_ken_table_aliases`, and :func:`datasets.get_ken_types` will be
   removed in the next release of skrub.
   :pr:`1546` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
+- Improved error messages when a DataOp is being sent to dispatched functions.
+  :pr:`1607` by :user:`Riccardo Cappuzzo<rcap107>`.
+- The accepted values for the parameter ``how`` of :meth:`DataOp.skb.apply` have
+  changed. The new values are ``"auto"`` (unchanged), ``"cols"`` to wrap the
+  transformer in :class:`ApplyToCols`, ``"frame"`` to wrap the transformer in
+  :class:`ApplyToFrame`, or ``"no_wrap"`` for no wrapping. The old values are
+  deprecated and will result in an error in a future release.
+  :pr:`1628` by :user:`Jérôme Dockès <jeromedockes>`.
+- The parameter ``splitter`` of :meth:`DataOp.skb.train_test_split` has been
+  renamed ``split_func``. :pr:`1630` by :user:`Jérôme Dockès <jeromedockes>`.
 
 - The objects ``tabular_learner`` and ``DropIfTooManyNulls`` were removed. Use
   :func:`tabular_pipeline` and :class:`DropUninformative` instead.
@@ -47,6 +99,9 @@ Bugfixes
 - DataOp hyperparameter search would raise an error when doing classification
   and using the ``scoring`` parameter, when the dataop contained no variables.
   Fixed in :pr:`1601` by :user:`Jérôme Dockès <jeromedockes>`.
+- :class:`SkrubLearner` used to do a prediction on the train set during
+  ``fit()``, this has been fixed.
+  :pr:`1610` by :user:`Jérôme Dockès <jeromedockes>`.
 - :class:`DataOp` would raise errors when containing subclasses of list, tuple
   or dict that cannot be initialized with an instance of the builtin type (such
   as classes created by ``collections.namedtuple``), this has been fixed.
@@ -54,6 +109,11 @@ Bugfixes
   (not into their subclasses). If you need the items evaluated (ie if they
   contain DataOps or Choices), store them in one of the builtin collections.
   :pr:`1612` by :user:`Jérôme Dockès <jeromedockes>`.
+- :meth:`SkrubLearner.report` with ``mode="fit"`` used to display the dataops
+  themselves, rather than their outputs, in the report. This has been fixed in
+  :pr:`1623` by :user:`Jérôme Dockès <jeromedockes>`.
+- Fixed a bug that happened when ``get_feature_names_out`` was called on instances
+  of the :class:`DatetimeEncoder`. :pr:`1622` by :user:`Riccardo Cappuzzo<rcap107>`.
 
 Release 0.6.1
 ===================
@@ -79,7 +139,7 @@ Highlights
   to train and tune machine learning models. Then, the DataOps plans can be exported
   as ``Learners`` (:class:`skrub.SkrubLearner`), standalone objects that can be
   used on new data. More detail about the DataOps can be found in the
-  :ref:`User guide <userguide_data_ops>` and in the
+  :ref:`User guide <user_guide_data_ops_index>` and in the
   :ref:`examples <data_ops_examples_ref>`.
 
 - The :class:`TableReport` has been improved with many new features. Series are
@@ -90,12 +150,12 @@ Highlights
 
 - :mod:`selectors`, :class:`ApplyToCols` and :class:`ApplyToFrame` are now available,
   providing utilities for selecting columns to which a transformer should be applied
-  in a flexible way. For more details, see the :ref:`User guide <userguide_selectors>`
-  and the :ref:`example <sphx_glr_auto_examples_10_apply_on_cols.py>`.
+  in a flexible way. For more details, see the :ref:`User guide <user_guide_selectors>`
+  and the :ref:`example <sphx_glr_auto_examples_09_apply_to_cols.py>`.
 
 - The :class:`SquashingScaler` has been added: it robustly rescales and smoothly
-  clips numerical columns, enabling more robust handling of numerical columns
-  with neural networks. See the :ref:`example <sphx_glr_auto_examples_11_squashing_scaler.py>`
+  clips numeric columns, enabling more robust handling of numeric columns
+  with neural networks. See the :ref:`example <sphx_glr_auto_examples_10_squashing_scaler.py>`
 
 New features
 ------------
@@ -141,8 +201,8 @@ New features
   in reports. :pr:`1470` by :user:`George S <georgescutelnicu>`.
 
 - Added the :class:`SquashingScaler`, a transformer that
-  robustly rescales and smoothly clips numerical columns,
-  enabling more robust handling of numerical columns
+  robustly rescales and smoothly clips numeric columns,
+  enabling more robust handling of numeric columns
   with neural networks. :pr:`1310` by :user:`Vincent Maladiere <Vincent-Maladiere>` and
   :user:`David Holzmüller <dholzmueller>`.
 
@@ -210,7 +270,7 @@ Changes
 
 - The :class:`TableReport` now supports Series in addition to Dataframes. :pr:`1420` by :user:`Vitor Pohlenz<vitorpohlenz>`.
 
-- The :class:`Cleaner` now exposes a parameter to convert numerical values to float32. :pr:`1440` by
+- The :class:`Cleaner` now exposes a parameter to convert numeric values to float32. :pr:`1440` by
   :user:`Riccardo Cappuzzo<rcap107>`.
 
 - The :class:`TableReport` now shows if columns are sorted. :pr:`1512` by :user:`Dea María Léon<DeaMariaLeon>`.
@@ -341,8 +401,7 @@ Release 0.4.1
 Changes
 -------
 
-* :class: `TableReport` has `write_html` method
-  :pr:`1190` by :user:`Mojdeh Rastgoo<mrastgoo>`.
+* :class:`TableReport` has `write_html` method. :pr:`1190` by :user:`Mojdeh Rastgoo<mrastgoo>`.
 
 * A new parameter ``verbose`` has been added to the :class:`TableReport` to toggle on or off the
   printing of progress information when a report is being generated.
@@ -810,7 +869,7 @@ Minor changes
   - `check_is_fitted` now looks at `"transformers_"` rather than `"columns_"`
   - the default of the `remainder` parameter in the docstring is now `"passthrough"`
     instead of `"drop"` to match the implementation.
-  - uint8 and int8 dtypes are now considered as numerical columns.
+  - uint8 and int8 dtypes are now considered as numeric columns.
 
 * Removed the leading "<" and trailing ">" symbols from KEN entities
   and types.
@@ -860,10 +919,10 @@ Dirty-cat release 0.4.1
 
 Major changes
 -------------
-* :func:`fuzzy_join` and :class:`FeatureAugmenter` can now join on numerical columns based on the euclidean distance.
+* :func:`fuzzy_join` and :class:`FeatureAugmenter` can now join on numeric columns based on the euclidean distance.
   :pr:`530` by :user:`Jovan Stojanovic <jovan-stojanovic>`
 
-* :func:`fuzzy_join` and :class:`FeatureAugmenter` can perform many-to-many joins on lists of numerical or string key columns.
+* :func:`fuzzy_join` and :class:`FeatureAugmenter` can perform many-to-many joins on lists of numeric or string key columns.
   :pr:`530` by :user:`Jovan Stojanovic <jovan-stojanovic>`
 
 * :func:`GapEncoder.transform` will not continue fitting of the instance anymore.
@@ -953,7 +1012,7 @@ Dirty-cat Release 0.3.0
 Major changes
 -------------
 
-* New encoder: :class:`DatetimeEncoder` can transform a datetime column into several numerical columns
+* New encoder: :class:`DatetimeEncoder` can transform a datetime column into several numeric columns
   (year, month, day, hour, minute, second, ...). It is now the default transformer used
   in the :class:`TableVectorizer` for datetime columns. :pr:`239` by :user:`Leo Grinsztajn <LeoGrin>`